requirements.txt
fastapi
uvicorn[standard]
qdrant-client
uniface
onnxruntime
opencv-python-headless
numpy
python-dotenv
scikit-image
python-multipart
tensorboard
easydict